# specs/openapi/qa-lab-api.yaml
openapi: 3.0.3
info:
  title: QA Lab API
  description: >-
    Minimal local test subject for Story 1.2. In-memory item resource with
    documented success and error behavior. Restarting the process resets all
    state. There is no reset endpoint. Automated schema validation arrives in
    Story 1.3.
  version: 0.1.0
servers:
  - url: http://127.0.0.1:3001
    description: Local default. Override the port with QA_LAB_API_PORT.
paths:
  /health:
    get:
      summary: Liveness check
      operationId: getHealth
      responses:
        "200":
          description: API is running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /items:
    get:
      summary: List stored items
      operationId: listItems
      responses:
        "200":
          description: Current in-memory items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemList"
    post:
      summary: Create an item
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItem"
            examples:
              notebook:
                value:
                  name: Notebook
                  quantity: 2
      responses:
        "201":
          description: Item created and stored in memory.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Invalid input type, missing data, or quantity outside 1..100.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /items/{id}:
    get:
      summary: Get one item by id
      operationId: getItemById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: item-0001
      responses:
        "200":
          description: The stored item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: No item stored under this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
components:
  schemas:
    Health:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
    Item:
      type: object
      required:
        - id
        - name
        - quantity
      properties:
        id:
          type: string
          example: item-0001
        name:
          type: string
          example: Notebook
        quantity:
          type: integer
          minimum: 1
          maximum: 100
          example: 2
    ItemList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    CreateItem:
      type: object
      required:
        - name
        - quantity
      additionalProperties: true
      properties:
        name:
          type: string
          example: Notebook
        quantity:
          type: integer
          minimum: 1
          maximum: 100
          example: 2
    ValidationError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Field 'quantity' must be between 1 and 100.
    NotFound:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Item 'item-9999' was not found.